+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Changed
+- Compatibility with newer versions of IntelliJ has been improved.
+- Support for versions older than 2022.1 has been removed. Version 1.\*.\* of Copy Commit Reference can still be used with any versions of IntelliJ between 2020.3 and 2023.2.
+
 ## [1.1.1] - 2023-08-10
 
 ### Fixed